azure_functions/requirements.txt
# Requirements for OC-P10 Recommender API
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
azure-functions
